Added vimscript function to import tw5 from vim
Inputs wiki text for multiple tiddlers from vim
directly into specified Tiddlywiki 5 node.js installation.

Advantages:
- NO MORE dragging and dropping json files to your wiki.
- NO MORE scrolling up and down to add captions, order, and tags to
          your tiddlers for the the tab macro.

Dependencies: jq must be installed and available on your path in bash.

Best Practices:
- Use an ssh-agent and an ssh public/private key to avoid typing your
  password everytime.
 (see attached sample upload document for how to set that up)
- Map the function call for the document you are working on (local
  machine) and host (remote server) to a function key in .vimrc.

More Detail:
1. jq reads the contents of the specified text file into a json file
   of tiddlers stored in a local temp file.

2. The temp json file is then uploaded to the Tiddlywiki5 host in
   the same temp directory as was created on the local host.

3. The Tiddlywiki 5 node executable is then called on the remote server
   and the json temp file tiddlers are loaded into Tiddlywiki 5.

4. Restart the specified Tiddlywiki 5 instance and load it up in your
   browser.  It will now contain the tiddlers you created.

Syntax:

Each tiddler (minus the first one) is seperated with `----` followed by
two line breaks.

Next the title of the tiddler is listed with an `^!` in front and a line
break.

On the folowing line, the order of said tab is specifed after a `%n`
where n is a number specfing the order of the tiddler in another
tiddlers' set of tabs, followed by a line break.

On the next line a `^;` is used to specify a caption for the tab on the
other tiddler, followed by a line break.

On the next line an `^&` specifies a list of tags for the tiddler (these
are used to group together groups of tiddlers, such as for a parent set
of tabs on another tiddler).  After this, specify two more line breaks,
and begin the tiddler's text content.